+++ routes/event-model.js
@@ -8,29 +8,29 @@ module.exports = {
     update,
 }
 
-function get(cal_id) {
+function get(id) {
     return (
         db('calendarEvents')
-            .where({cal_id})
-            // .select('calendarId', 'eventsId')
+            .where({id})
+            .select('calendarId', 'eventsId')
             // .join('events', 'eventsId', 'events.id')
             // .select('eventName', 'eventInfo')
     )
 }
 
-function getById(cal_id, event_id) {
+function getById(id, event_id) {
     return (
         db('calendarEvents')
-            .where({cal_id, event_id})
+            .where({id, event_id})
             .join('events', 'eventsId', 'events.id')
             .select('eventName', 'eventInfo')
     )
 }
 
-function add(cal_id, event) {
+function add(id, event) {
     return (
         db('calendarEvents')
-            .where({cal_id})
+            .where({id})
             .insert(event)
     )
 }
